Добавил webp, валидация трёх страниц, gulpfile - старый для удобства правок
